Add script to convert binary region_mask to netCDF
Gokhan suggests using the Atlantic Region Mask as a way to test the SST forcing
setup (alpha = 0 in Atlantic Region and 1 elsewhere, except for a transition
region to the north and south. The region_mask_to_netCDF.ncl script converts
the binary file containing the region mask (in inputdata/) to netCDF, which
makes it much easier to read / visualize.