Update _FillValue in a couple scripts
Changed alpha from a float to a double, and updated the fill value to
be -1 (since 0 <= alpha <= 1 everywhere in the ocean). Also updated
the fill value in the SST dataset to be 273.15 (not really necessary
or useful, except I can never remember how precise CESM is in conversion
between degrees C and degrees K).

Also updated my notes because I tested a couple of B1D compsets to make
sure everything went through the coupler correctly. It did, I think.